ciri crypto module
