+++ ciri-crypto/.rspec
@@ -1,0 +1,3 @@
+--format documentation
+--color
+--require spec_helper
