+++ ciri-crypto/lib/ciri/crypto/version.rb
@@ -1,0 +1,5 @@
+module Ciri
+  module Crypto
+    VERSION = "0.1.0"
+  end
+end
